+++ prisma/schema.prisma
@@ -29,18 +29,19 @@ model Secretaria {
 }
 
 model Medico {
-  usuario_id       String            @id
-  tipo_matricula   TipoMatriculaEnum
-  numero_matricula String            @unique
-  numero_telefono  String
-  ubicacion_id     String            @unique
-  especialidad_id  String
-  descripcion      String?
-  duracion_cita    Int?
-  citas            Cita[]            @relation("MedicoCitas")
-  especialidad     Especialidad      @relation(fields: [especialidad_id], references: [id])
-  ubicacion        Ubicacion         @relation(fields: [ubicacion_id], references: [id])
-  usuario          Usuario           @relation(fields: [usuario_id], references: [id])
+  usuario_id        String              @id
+  tipo_matricula    TipoMatriculaEnum
+  numero_matricula  String              @unique
+  numero_telefono   String
+  ubicacion_id      String              @unique
+  especialidad_id   String
+  descripcion       String?
+  duracion_cita     Int?
+  citas             Cita[]              @relation("MedicoCitas")
+  IntervaloAtencion IntervaloAtencion[]
+  especialidad      Especialidad        @relation(fields: [especialidad_id], references: [id])
+  ubicacion         Ubicacion           @relation(fields: [ubicacion_id], references: [id])
+  usuario           Usuario             @relation(fields: [usuario_id], references: [id])
 }
 
 model Paciente {
@@ -111,6 +112,15 @@ model Cita {
   paciente    Paciente @relation("PacienteCitas", fields: [paciente_id], references: [usuario_id])
 }
 
+model IntervaloAtencion {
+  id         String        @id
+  diaSemana  DiaSemanaEnum
+  horaInicio DateTime
+  horaFin    DateTime
+  medico_id  String
+  Medico     Medico        @relation(fields: [medico_id], references: [usuario_id])
+}
+
 enum TipoDocumentoEnum {
   CEDULA_IDENTIDAD
   DOCUMENTO_NACIONAL_IDENTIDAD
@@ -122,4 +132,12 @@ enum TipoDocumentoEnum {
 enum TipoMatriculaEnum {
   nacional
   provincial
+}
+
+enum DiaSemanaEnum {
+  LUNES
+  MARTES
+  MIERCOLES
+  JUEVES
+  VIERNES
 }
